opt: remove helper due to impact on coverage
