+++ internal/infrastructure/connector/az_blob_connector_integration_test.go
@@ -5,6 +5,9 @@ package connector
 
 import (
 	"context"
+	"fmt"
+	"mime/multipart"
+	"os"
 	"testing"
 
 	"crypto_vault_service/internal/infrastructure/logger"
@@ -15,6 +18,32 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
+
+// Helper function to create test files
+func CreateTestFile(fileName string, content []byte) error {
+	err := os.WriteFile(fileName, content, 0600)
+	if err != nil {
+		return fmt.Errorf("failed to create test file: %w", err)
+	}
+	return nil
+}
+
+// Helper function to create a test file and form
+func CreateTestFileAndForm(t *testing.T, fileName string, fileContent []byte) (*multipart.Form, error) {
+	err := CreateTestFile(fileName, fileContent)
+	require.NoError(t, err)
+
+	t.Cleanup(func() {
+		if err := os.Remove(fileName); err != nil {
+			t.Logf("failed to remove temporary file %s: %v", fileName, err)
+		}
+	})
+
+	form, err := utils.CreateForm(fileContent, fileName)
+	require.NoError(t, err)
+
+	return form, nil
+}
 
 // AzureBlobConnectorTest encapsulates common logic for tests
 type AzureBlobConnectorTest struct {
@@ -51,7 +80,7 @@ func TestAzureBlobConnector_Upload(t *testing.T) {
 
 	testFileContent := []byte("This is test file content")
 	testFileName := "testfile.txt"
-	form, err := utils.CreateTestFileAndForm(t, testFileName, testFileContent)
+	form, err := CreateTestFileAndForm(t, testFileName, testFileContent)
 	require.NoError(t, err)
 
 	userId := uuid.New().String()
@@ -80,7 +109,7 @@ func TestAzureBlobConnector_Download(t *testing.T) {
 
 	testFileContent := []byte("This is test file content")
 	testFileName := "testfile.pem"
-	form, err := utils.CreateTestFileAndForm(t, testFileName, testFileContent)
+	form, err := CreateTestFileAndForm(t, testFileName, testFileContent)
 	require.NoError(t, err)
 
 	userId := uuid.New().String()
@@ -108,7 +137,7 @@ func TestAzureBlobConnector_Delete(t *testing.T) {
 
 	testFileContent := []byte("This is test file content")
 	testFileName := "testfile.pem"
-	form, err := utils.CreateTestFileAndForm(t, testFileName, testFileContent)
+	form, err := CreateTestFileAndForm(t, testFileName, testFileContent)
 	require.NoError(t, err)
 
 	userId := uuid.New().String()
